+++ pom.xml
@@ -7,6 +7,10 @@
     <groupId>groupId</groupId>
     <artifactId>libraryCatalog</artifactId>
     <version>1.0-SNAPSHOT</version>
+    <properties>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+    </properties>
 
-    
+
 </project>
