+++ src/main/java/ArtAlbum.java
@@ -1,5 +1,3 @@
-package libraryCatalog;
-
 /** here we made a subclass for class Books
  * it contains a variable type String for page quality
  */
